requirements.txt
discord.py
python-dotenv
docker
psutil
